Add GET request handler
Why these changes are being introduced:
The function needs to check for a `challenge` parameter in incoming GET
requests and respond with that parameter's value. This is how Alma
makes the initial webhook connection during configuration.

How this addresses that need:
* Adds functionality and tests to check the request method, respond
  appropriately to GET requests, and provide error responses to other
  request methods and incorrectly supplied parameters
* Updates the README local development instructions to reflect changes
* Also updates the Dockerfile and Makefile to handle pipenv's creation
  of a Requirements.txt file within the container

Side effects of this change:
GET requests to the URL function must now include a `challenge`
parameter to receive a success response.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-66