Add body checks to identify POD export job
Why these changes are being introduced:
Currently this webhook should only trigger an action if the POST
request is sent for a POD export job that completed successfully and
exported records to S3. We need to trigger the step function in that
case (actual step function trigger will be handled in a future commit)
and log/send an appropriate response for all other cases (not a
JOB_END, not a POD export job, POD export job failed, or POD export job
completed successfully with no records exported).

How this addresses that need:
* Updates lambda_handler function to check for a non-JOB_END job
  and send an error response because we do not currently expect Alma
  to be configured to send any other type of webhook to this function.
* Updates lambda_handler function to check for a POD export job and
  call the pod export job handler, otherwise returns default "no action
  taken" response.
* Adds handle_pod_export_webhook function that checks for a successful
  job completion with exported records and logs/returns the step
  function trigger if so (currently just a placeholder), or returns a
  "no action taken" response. Also logs a warning suggesting further
  investigation in case of export job failure.
* Adds tests for all new functionality.
* Minor update to valid_signature function and tests to reflect
  expected webhook body format.

Side effects of this change:
This lambda now requires a new env variable to run successfully,
ALMA_POD_EXPORT_JOB_NAME.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-73