Configure Automated Deploys
Why these changes are being introduced:
The changes here update the automated deploy process for the Lambda
function using GitHub Actions. It also updates the Makefile for the
local developer for easily deploying test builds in Dev1 (and provides
a framework for emergency builds/deploys in Stage-Workloads).

How this addresses that need:
* Update Makefile with Terraform outputs from mitlib-tf-workloads-ecr
* Update dev-build.yml caller workflow with Terraform outputs from
mitlib-tf-workloads-ecr
* Update stage-build.yml caller workflow with Terraform outputs from
miblib-tf-workloads-ecr
* Create prod-promote.yml caller workflow with Terraform outputs from
mitlib-tf-workloads-ecr

Side effects of this change:
This repo is now fully configured for GHA-based automated deploys to
AWS.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ENSY-79