Fix POD job filename date bug
Why these changes are being introduced:
It was unclear from Alma documentation whether the date in publishing
profile file name exports used the job start time or end time if the job
went over two days. We assumed start time, but it turns out it uses the
end time so we need to pull that date from the webhook to generate the
file name prefix to send to the step function.

How this addresses that need:
* Replaces "start_time" with "end_time" in file name generation for pod
  export job handling.
* Updates relevant test to match.
* Also adds vscode files to gitignore.