Fix Sentry integration
Why these changes are being introduced:
Lambda functions require a special sentry integration option which we
had not previously used so the integration was broken.

How this addresses that need:
* Adds Sentry lambda integration to webhook module and moves sentry init
  to top of module as required instead of within the lambda_handler
  function.
* Adds check for required WORKSPACE env in lambda_handler function.
* Adds tests for new functionality.
* Updates README to include local testing with Docker.
* Removes duplicate line in Dockerfile.

Side effects of this change:
Exceptions should now be sent to Sentry as expected in all AWS envs.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-526