Check env match in job name
Why these changes are being introduced:
For complicated reasons having to do with Alma configuration and the
dev/stage/prod pipeline, we need to check that the job name matches the
environment this webhook is running in before triggering an action. See
README updates for more detail.

How this addresses that need:
* Adds a check in the job end handler to confirm that the job name
  contains the webhook runtime environment. If not, does not trigger the
  relevant action.
* Updates tests and README to reflect changes.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-176