job names include aws env in SSM
Why these changes are being introduced:
We changed some configurations in AWS so that the job names are all
stored in SSM. Because of this we can include the AWS environment in
the job name, which should be the complete job name for POD and Bursar
jobs, TIMDEX is an exception. The TIMDEX job name should include the
AWS environment, but is still technically a prefix because TIMDEX job
names end with the type (full or daily), but we don't want to have two
separate environment variables for each of those job types.

How this addresses that need:
* Updated readme
* Changed the name of the bursar job name env variable
* Reordered the check for matching environments to come first
* updated the tests to use the new variable name
* updated the tests to inclue the environment name in test environment
variables