update bursar step function trigger
Why are these changes being introduced:
The bursar step function call was throwing an error because the job
name being passed was not unique.

We also changed how we are passing the target and source s3 prefix to
the bursar lambda, so passing the Alma job name to the bursar step function
is no longer required.

relevant tickets:
https://mitlibraries.atlassian.net/browse/ENSY-178

https://mitlibraries.atlassian.net/browse/ENSY-172

How this addresses that need:
* add a timestamp to the step function execution name parameter so it is
unique.

* bursar transfer lambda only needs the
job ID, not the job name.